movement update small
adding boundaries and rules for movement so that people cannot walk through rocks and trees and such.
made a specific list so swimmers can cross water.
so far it stopped everyone from moving but it doesnt break.
i need to find a method that allows me to change values to strings.